Added implosives
Hayes (2009) mentions the feature [implosive] that is positive for implosive sounds but he doesn't include them in his feature list. I have made a compromise by leaving the features as they are and making implosive sounds have the feature +constr_gl alongside +voice. This may not be strictly phonetically detailed enough but it's similar to the issue of ATR and tense being represented by the same feature in Hayes' features. There is some relationship between creaky voice and implosives "Like the contrast between breathy and modal voiced among obstruents, contrasts between creaky and modal voice are also relatively rare in obstruents, though Hausa and certain other Chadic languages make such a contrast for stops. The creaky stops in these languages are implosives and involve larynx lowering as well as a creaky voice quality" Gordon and Ladefoged (2001)